3.2-crud/stocks_products/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.3
django-bootstrap4==23.2
django-debug-toolbar==4.1.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-pagination==1.1
Pillow==10.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.7
pytz==2023.3
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3